fix(ci): resolve AudioPlayer glob collision and redundant cast
Both CI jobs have failed on every push since 2026-06-14.

test: `widgets.rs` does `use bevy::prelude::*` (private) alongside
`pub use audio_player::*`. Both carry an `AudioPlayer`, and a private glob
beats a public one, so `renzora_ember::widgets::AudioPlayer` resolved to
bevy's and `renzora_hub` failed with E0603. An explicit re-export shadows
both globs. The collision only materialises under the feature unification
`cargo test --workspace` produces, which is why `--profile dist` builds
were unaffected.

clippy: `max_storage_buffer_binding_size` is already u64, so `as u64`
tripped `unnecessary_cast` under `-D warnings`.

+++ crates/renzora_gaussian_splatting/src/lib.rs
@@ -88,7 +88,7 @@ fn reject_oversized_clouds(
     let limits = device.limits();
     let max_bytes = limits
         .max_buffer_size
-        .min(limits.max_storage_buffer_binding_size as u64);
+        .min(limits.max_storage_buffer_binding_size);
     const WORST_BYTES_PER_SPLAT: u64 = 192;
 
     for event in events.read() {
